Redesign video player layout to be more flexible
