+++ app/shared/commonMain/ui/subject/episode/EpisodeVideo.kt
@@ -1,20 +1,29 @@
 package me.him188.ani.app.ui.subject.episode
 
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.animation.core.tween
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.material3.LocalContentColor
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ProvideTextStyle
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -22,21 +31,121 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.delay
 import me.him188.ani.app.platform.isInLandscapeMode
 import me.him188.ani.app.ui.foundation.TopAppBarGoBackButton
 import me.him188.ani.app.ui.theme.aniDarkColorTheme
-import me.him188.ani.app.ui.theme.looming
+import me.him188.ani.app.ui.theme.slightlyWeaken
 import me.him188.ani.app.videoplayer.PlayerController
 import me.him188.ani.app.videoplayer.VideoPlayerView
-import me.him188.ani.app.videoplayer.ui.PlayerControllerOverlay
-import me.him188.ani.app.videoplayer.ui.PlayerControllerOverlayTopBar
+import me.him188.ani.app.videoplayer.ui.PlayerNavigationBar
 import me.him188.ani.app.videoplayer.ui.PlayerProgressController
 import me.him188.ani.app.videoplayer.ui.VideoLoadingIndicator
 import moe.tlaster.precompose.flow.collectAsStateWithLifecycle
 import kotlin.time.Duration.Companion.seconds
+
+
+@Composable
+internal fun VideoScaffold(
+    controllersVisible: Boolean = true,
+    onControllersVisibleChange: (Boolean) -> Unit,
+    topBar: @Composable RowScope.() -> Unit = {},
+    video: @Composable BoxScope.() -> Unit = {},
+    danmakuHost: @Composable BoxScope.() -> Unit = {},
+    floatingMessage: @Composable BoxScope.() -> Unit = {},
+    bottomBar: @Composable RowScope.() -> Unit = {},
+    modifier: Modifier = Modifier,
+    isFullscreen: Boolean = isInLandscapeMode(),
+) {
+    BoxWithConstraints(
+        modifier.then(if (isFullscreen) Modifier.fillMaxHeight() else Modifier.fillMaxWidth()),
+        contentAlignment = Alignment.Center
+    ) { // 16:9 box
+        Box(
+            Modifier
+                .then(
+                    if (isFullscreen) {
+                        Modifier.fillMaxSize()
+                    } else {
+                        Modifier.fillMaxWidth().height(maxWidth * 9 / 16) // 16:9 box
+                    }
+                )
+        ) {
+            Box(Modifier
+                .clickable(
+                    remember { MutableInteractionSource() },
+                    indication = null,
+                    onClick = {
+                        onControllersVisibleChange(!controllersVisible)
+                    }
+                )
+                .matchParentSize()
+            ) {
+                video()
+                Box(Modifier.matchParentSize()) // 防止点击事件传播到 video 里
+            }
+
+            Column(Modifier.fillMaxSize().background(Color.Transparent)) {
+                val controllerBackground = aniDarkColorTheme().background.copy(0.8f)
+
+                // 顶部控制栏: 返回键, 标题, 设置
+                AnimatedVisibility(
+                    visible = controllersVisible,
+                    enter = fadeIn(),
+                    exit = fadeOut(),
+                ) {
+                    Row(
+                        Modifier
+                            .fillMaxWidth()
+                            .background(color = controllerBackground)
+                            .statusBarsPadding()
+                            .padding(top = 12.dp)
+                            .height(32.dp),
+                        verticalAlignment = Alignment.CenterVertically,
+                    ) {
+                        CompositionLocalProvider(LocalContentColor provides aniDarkColorTheme().onBackground) {
+                            topBar()
+                        }
+                    }
+                }
+
+                // 弹幕
+                Box(Modifier.weight(1f, fill = true).padding(all = 8.dp)) {
+                    danmakuHost()
+                }
+
+                // 底部控制栏: 播放/暂停, 进度条, 切换全屏
+                AnimatedVisibility(
+                    visible = controllersVisible,
+                    enter = fadeIn(),
+                    exit = fadeOut(),
+                ) {
+                    Row(
+                        Modifier
+                            .fillMaxWidth()
+                            .background(color = controllerBackground)
+                            .height(40.dp),
+                        verticalAlignment = Alignment.CenterVertically,
+                    ) {
+                        CompositionLocalProvider(LocalContentColor provides Color.White) {
+                            bottomBar()
+                        }
+                    }
+                }
+            }
+
+            Box(Modifier.matchParentSize(), contentAlignment = Alignment.Center) {
+                ProvideTextStyle(MaterialTheme.typography.labelSmall) {
+                    CompositionLocalProvider(LocalContentColor provides aniDarkColorTheme().onBackground.slightlyWeaken()) {
+                        floatingMessage()
+                    }
+                }
+            }
+        }
+    }
+}
 
 /**
  * 剧集详情页面顶部的视频控件.
@@ -53,100 +162,70 @@ internal fun EpisodeVideo(
 ) {
     val (controllerVisible, setControllerVisible) = remember { mutableStateOf(true) }
 
-    val controllerAlpha by animateFloatAsState(if (controllerVisible) 1f else 0f, tween())
-//    LaunchedEffect(controllerVisible) {
-//        // 2 秒后隐藏 TopAppBar
-//        if (controllerVisible) {
-//            launch {
-//                delay(2.seconds)
-//                setControllerVisible(false)
-//            }
-//        }
-//    }
-
-    BoxWithConstraints(
-        modifier.then(if (isFullscreen) Modifier.fillMaxHeight() else Modifier.fillMaxWidth()),
-        contentAlignment = Alignment.Center
-    ) {
-        Box(
-            Modifier
-                .then(
-                    if (isFullscreen) {
-                        Modifier.fillMaxSize()
-//                    Modifier.fillMaxHeight().width(maxHeight * 16 / 9)
-                    } else {
-                        Modifier.fillMaxWidth()
-                            .height(maxWidth * 9 / 16)
-                    }
-                ).clickable(
-                    remember { MutableInteractionSource() },
-                    indication = null,
-                    onClick = {
-                        setControllerVisible(!controllerVisible)
-                    }
-                )
-        ) { // 16:9 box
-            VideoPlayerView(playerController, Modifier.matchParentSize())
-
-            PlayerControllerOverlay(
-                topBar = {
-                    PlayerControllerOverlayTopBar(
-                        startActions = {
-                            TopAppBarGoBackButton(onClickGoBack)
-                        },
-                        Modifier.alpha(controllerAlpha)
-                            .background(color = aniDarkColorTheme().background.copy(0.8f))
-                            .statusBarsPadding()
-                            .padding(top = 12.dp)
-                            .height(24.dp),
-                    )
+    VideoScaffold(
+        controllersVisible = controllerVisible,
+        onControllersVisibleChange = setControllerVisible,
+        topBar = {
+            PlayerNavigationBar(
+                actions = {
+                    TopAppBarGoBackButton(onClickGoBack)
                 },
-                floatingBox = {
-                    val isBuffering by playerController.isBuffering.collectAsStateWithLifecycle(true)
-                    if (isBuffering) {
-                        var loadedTooLong by remember { mutableStateOf(false) }
-                        VideoLoadingIndicator(
-                            showProgress = videoSourceSelected,
-                            text = {
-                                when {
-                                    !videoSourceSelected -> {
-                                        Text("请选择数据源")
-                                    }
-
-                                    !videoReady -> {
-                                        Text("正在下载种子")
-                                    }
-
-                                    loadedTooLong -> {
-                                        Text("资源较慢, 正在努力缓冲")
-                                    }
-
-                                    else -> {
-                                        //                                val speed by video.downloadRate.collectAsStateWithLifecycle(null)
-                                        Text("正在缓冲")
-                                        LaunchedEffect(true) {
-                                            delay(10.seconds)
-                                            loadedTooLong = true
-                                        }
-                                    }
-                                }
-                            }
-                        )
-                    }
-                },
-                bottomBar = {
-                    PlayerProgressController(
-                        controller = playerController,
-                        onClickFullScreen = onClickFullScreen,
-                        modifier = Modifier
-                            .alpha(controllerAlpha)
-                            .background(color = aniDarkColorTheme().background.looming())
-                            .padding(vertical = 4.dp)
-                            .fillMaxWidth()
-                    )
-                },
-                Modifier.matchParentSize()
             )
-        }
+        },
+        video = {
+            VideoPlayerView(playerController, Modifier.matchParentSize())
+        },
+        bottomBar = {
+            PlayerProgressController(
+                controller = playerController,
+                onClickFullScreen = onClickFullScreen,
+            )
+        },
+        floatingMessage = {
+            EpisodeVideoLoadingIndicator(playerController, videoSourceSelected, videoReady)
+        },
+        modifier = modifier,
+        isFullscreen = isFullscreen
+    )
+}
+
+@Composable
+private fun EpisodeVideoLoadingIndicator(
+    playerController: PlayerController,
+    videoSourceSelected: Boolean,
+    videoReady: Boolean,
+    modifier: Modifier = Modifier,
+) {
+    val isBuffering by playerController.isBuffering.collectAsStateWithLifecycle(true)
+    if (isBuffering) {
+        var loadedTooLong by remember { mutableStateOf(false) }
+        VideoLoadingIndicator(
+            showProgress = videoSourceSelected,
+            text = {
+                when {
+                    !videoSourceSelected -> {
+                        Text("请选择数据源")
+                    }
+
+                    !videoReady -> {
+                        Text("正在下载种子")
+                    }
+
+                    loadedTooLong -> {
+                        Text("资源较慢, 正在努力缓冲")
+                    }
+
+                    else -> {
+                        //                                val speed by video.downloadRate.collectAsStateWithLifecycle(null)
+                        Text("正在缓冲")
+                        LaunchedEffect(true) {
+                            delay(10.seconds)
+                            loadedTooLong = true
+                        }
+                    }
+                }
+            },
+            modifier,
+        )
     }
 }
